菜单功能修改

+++ leon-web-common/src/main/java/com/mawujun/fun/FunService.java
@@ -41,7 +41,7 @@ public class FunService extends BaseRepository<Fun, String> {
 		
 
 		//menuItemService.delete(fun);
-		int menuItemCount=menuItemService.queryCount(Cnd.select().andEquals("fun.id", entity.getId()));
+		long menuItemCount=menuItemService.queryCount(Cnd.select().andEquals("fun.id", entity.getId()));
 		if(menuItemCount>0){
 			throw new BusinessException("有菜单挂钩，不能删除。",WebCommonExceptionCode3.EXISTS_CHILDREN);
 		}
